Change cost and cardinality metrics to always return Double

+++ community/cypher/cypher-compiler-2.1/src/main/scala/org/neo4j/cypher/internal/compiler/v2_1/planner/logical/Metrics.scala
@@ -28,11 +28,11 @@ object Metrics {
   // This metric calculates how expensive executing a logical plan is.
   // (e.g. by looking at cardinality, expression selectivity and taking into account the effort
   // required to execute a step)
-  type CostModel = LogicalPlan => Int
+  type CostModel = LogicalPlan => Double
 
   // This metric estimates how many rows of data a logical plan produces
   // (e.g. by asking the database for heuristics)
-  type CardinalityEstimator = LogicalPlan => Int
+  type CardinalityEstimator = LogicalPlan => Double
 
   // This metric estimates the selectivity of an expression
   // (e.g. by algebraic analysis or using heuristics)
